+++ src/controllers/OrgController.ts
@@ -1,5 +1,5 @@
- import { Request, Response } from "express";
-import { OrgService } from "../services/OrgService";
+import { Request, Response, NextFunction } from "express";
+import { OrgService } from "../services/organization.services";
 
 export class OrgController {
   private orgService: OrgService;
@@ -7,11 +7,36 @@ export class OrgController {
     this.orgService = new OrgService();
   }
 
+  async createOrganisation(req: Request, res: Response, next: NextFunction) {
+    try {
+      const payload = req.body;
+      const user = req.user;
+      const userId = user.id;
+
+      const organisationService = new OrgService();
+      const newOrganisation = await organisationService.createOrganisation(
+        payload,
+        userId
+      );
+
+      const respObj = {
+        status: "success",
+        message: "organisation created successfully",
+        data: newOrganisation,
+        status_code: 201,
+      };
+
+      return res.status(201).json(respObj);
+    } catch (error) {
+      return next(error);
+    }
+  }
+
   async removeUser(req: Request, res: Response) {
     try {
       const user = await this.orgService.removeUser(
         req.params.org_id,
-        req.params.user_id,
+        req.params.user_id
       );
       if (!user) {
         return res.status(404).json({
