+++ src/services/organization.services.ts
@@ -1,0 +1,86 @@
+import { Organization, User, UserOrganization } from "../models";
+import { UserRole } from "../enums/userRoles";
+import AppDataSource from "../data-source";
+import { HttpError } from "../middleware";
+import {
+  IOrgService,
+  IUserService,
+  ICreateOrganisation,
+  IOrganisationService,
+} from "../types";
+
+export class OrgService implements IOrgService {
+  public async createOrganisation(
+    payload: ICreateOrganisation,
+    userId: string
+  ): Promise<{
+    newOrganisation: Partial<Organization>;
+  }> {
+    try {
+      const organisation = new Organization();
+      organisation.owner_id = userId;
+      Object.assign(organisation, payload);
+
+      const newOrganisation = await AppDataSource.manager.save(organisation);
+
+      const userOrganization = new UserOrganization();
+      userOrganization.userId = userId;
+      userOrganization.organizationId = newOrganisation.id;
+      userOrganization.role = UserRole.ADMIN;
+
+      await AppDataSource.manager.save(userOrganization);
+
+      const user = await AppDataSource.getRepository(User).findOne({
+        where: { id: userId },
+        relations: ["userOrganizations", "userOrganizations.organization"],
+      });
+
+      user.userOrganizations.push(userOrganization);
+      await AppDataSource.getRepository(User).save(user);
+
+      return { newOrganisation };
+    } catch (error) {
+      throw new HttpError(error.status || 500, error.message || error);
+    }
+  }
+
+  public async removeUser(
+    org_id: string,
+    user_id: string
+  ): Promise<User | null> {
+    const userRepository = AppDataSource.getRepository(User);
+    const organizationRepository = AppDataSource.getRepository(Organization);
+
+    const user = await userRepository.findOne({
+      where: { id: user_id },
+      relations: ["organizations"],
+    });
+    if (!user) {
+      return null;
+    }
+
+    const organization = await organizationRepository.findOne({
+      where: { id: org_id },
+      relations: ["users"],
+    });
+    if (!organization) {
+      return null;
+    }
+
+    // Check if the user is part of the organization
+    const userInOrganization = organization.users.some(
+      (user) => user.id === user_id
+    );
+    if (!userInOrganization) {
+      return null;
+    }
+
+    // Remove the user from the organization
+    organization.users = organization.users.filter(
+      (user) => user.id !== user_id
+    );
+    await organizationRepository.save(organization);
+
+    return user;
+  }
+}
