+++ src/types/index.d.ts
@@ -5,9 +5,7 @@ export interface IUserService {
   getAllUsers(): Promise<User[]>;
 }
 
-export interface IOrgService {
-  removeUser(org_id: string, user_id: string): Promise<User | null>;
-}
+export interface IOrgService {}
 
 export interface IRole {
   role: "super_admin" | "admin" | "user";
@@ -44,7 +42,11 @@ export interface ICreateOrganisation {
 }
 
 export interface IOrganisationService {
-  createOrganisation(payload: ICreateOrganisation, userId: string): Promise<unknown>;
+  createOrganisation(
+    payload: ICreateOrganisation,
+    userId: string
+  ): Promise<unknown>;
+  removeUser(org_id: string, user_id: string): Promise<User | null>;
 }
 
 declare module "express-serve-static-core" {
